Refactored some of the Game->Keyboard->Command-.EnvironmentIterator->EnvironmentLoader communication, much more flexible now.
Made it so that the EnvironmentLoader just gets how ever many blocks that are needed, gives them the correct position, and draws them.  It shouldn't have to deal with the iterator; just tells the block which texture it should be according to the XML file.   It updates the block in case the texture changes.